refactor: project restructure based on postgres schema style
