--- src/support_sphere_db/src/ssdb/models/__init__.py
@@ -1,6 +1,0 @@
-from ssdb.models.user import User
-from ssdb.models.user_profile import UserProfile
-
-# New models created should be exposed by adding to __all__. This is used by SQLModel.metadata
-# https://sqlmodel.tiangolo.com/tutorial/create-db-and-table/#sqlmodel-metadata-order-matters
-__all__ = ['User', 'UserProfile']
