+++ src/support_sphere_db/src/ssdb/models/auth/__init__.py
@@ -1,0 +1,6 @@
+from ssdb.models.auth.user import User
+
+
+# New models created should be exposed by adding to __all__. This is used by SQLModel.metadata
+# https://sqlmodel.tiangolo.com/tutorial/create-db-and-table/#sqlmodel-metadata-order-matters
+__all__ = ['User']
