--- src/support_sphere_db/src/ssdb/repository/__init__.py
@@ -1,4 +1,0 @@
-from ssdb.repository.user_repository import UserRepository
-from ssdb.repository.user_profile_repository import UserProfileRepository
-
-__all__ = ['UserRepository', 'UserProfileRepository']
